+++ gradio/_vendor/aiofiles/__init__.py
@@ -1,0 +1,23 @@
+"""Utilities for asyncio-friendly file handling."""
+
+from . import tempfile
+from .threadpool import (
+    open,
+    stderr,
+    stderr_bytes,
+    stdin,
+    stdin_bytes,
+    stdout,
+    stdout_bytes,
+)
+
+__all__ = [
+    "open",
+    "tempfile",
+    "stdin",
+    "stdout",
+    "stderr",
+    "stdin_bytes",
+    "stdout_bytes",
+    "stderr_bytes",
+]