+++ js/atoms/src/BlockTitle.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-	import { default as Info } from "./Info.svelte";
+	import Info from "./Info.svelte";
+
 	export let show_label = true;
 	export let info: string | undefined = undefined;
 	export let rtl = false;