+++ js/dropdown/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/dropdown
 
+## 0.11.7
+
+### Dependency updates
+
+- @gradio/utils@0.12.2
+- @gradio/atoms@0.23.0
+- @gradio/statustracker@0.13.1
+
 ## 0.11.6
 
 ### Fixes